버그 수정 및 변경 사항
[+] 플레이어가 피격당했다는 것을 다른 클라이언트로 송신
[/] 플레이어가 움직이면서 죽었을 때 죽은 상태에서 움직이던 점 수정
[/] 몬스터의 타겟이 죽었을 경우 타겟을 변경하도록 변경
[/] 다른 사람이 쏜 총알은 더 큰 메쉬를 사용
[/] 스킬 게이지를 정수형에서 실수형으로 변경

+++ 소스/Client/player.cpp
@@ -926,9 +926,9 @@ void Player::SetGunShadowShader(const shared_ptr<Shader>& shadowShader)
 	m_gunShadowShader = shadowShader;
 }
 
-void Player::SetSkillGage(INT value)
+void Player::SetSkillGage(FLOAT value)
 {
-	m_skillGage = min(100, value);
+	m_skillGage = min(100.0f, value);
 }
 
 void Player::AddMaxHp(INT hp)
@@ -1018,7 +1018,7 @@ INT Player::GetMaxBulletCount() const
 	return m_maxBulletCount;
 }
 
-INT Player::GetSkillGage() const
+FLOAT Player::GetSkillGage() const
 {
 	return m_skillGage;
 }
@@ -1213,10 +1213,11 @@ void Player::UpdateSkill(FLOAT deltaTime)
 	}
 	else
 	{
+		// timerLimit초가 지나면 스킬게이지가 1씩 증가한다.
 		timerLimit = 1.0f;
 		if (m_skillGageTimer >= timerLimit)
 		{
-			SetSkillGage(m_skillGage + 1);
+			SetSkillGage(m_skillGage + 1.0f);
 			m_skillGageTimer -= timerLimit;
 		}
 	}
